+++ tests/testthat/test-get_tb_burden.R
@@ -29,6 +29,17 @@ test_that("TB burden has at least the expected number of entries", {
   expect_true(exp_nrows <= nrows_tb_data)
 })
 
+test_that("TB burden has no columns that are misjoined", {
+  mdr_tb_data <- get_tb_burden(download_data = TRUE,
+                               save = TRUE,
+                               add_mdr_data = TRUE)
+  
+  expect_equal(ncol(dplyr::select(mdr_tb_data, 
+                                 dplyr::ends_with(".x"), 
+                                 dplyr::ends_with(".y"))), 
+              c(as.integer(0)))
+})
+
 test_that("TB burden data is the same when downloaded using utils::read.csv", {
   skip_on_cran()
   expect_equal(tb_data, get_tb_burden(download_data = TRUE,