+++ NEWS.md
@@ -9,6 +9,8 @@
 * Improved data cleaning, converting `Inf` and `NaN` values to `NA` when the data is read in.
 * Added `pgknet` report.
 * Improved test robustness and scope
+* Added `vdiffr` to test plots when not on CRAN.
+* Fixed bug for `map_tb_burden` which was adding duplicate variables which caused map build to fail.
 
 # getTBinR 0.5.3.0