+++ NEWS.md
@@ -1,3 +1,11 @@
+# getTBinR 0.5.5
+
+## Feature updates
+
+
+## Package updates
+
+
 # getTBinR 0.5.4
 
 ## Feature updates